Moved to longs...

+++ cbits/help.c
@@ -1,13 +1,13 @@
 #include "help.h"
 
 void gcry_ac_io_init_readable_string(gcry_ac_io_t *ac_io,
-    unsigned char * str, size_t size)
+    unsigned char * str, unsigned long size)
 {
     gcry_ac_io_init(ac_io, GCRY_AC_IO_READABLE, GCRY_AC_IO_STRING, str, size);
 }
 
 void gcry_ac_io_init_writable_string(gcry_ac_io_t *ac_io,
-    unsigned char ** str, size_t * size)
+    unsigned char ** str, unsigned long * size)
 {
     gcry_ac_io_init(ac_io, GCRY_AC_IO_WRITABLE, GCRY_AC_IO_STRING, str, size);
 }
