+++ cbits/help.h
@@ -4,10 +4,10 @@
 #include <gcrypt.h>
 
 void gcry_ac_io_init_readable_string(gcry_ac_io_t *ac_io,
-    unsigned char * str, size_t size);
+    unsigned char * str, unsigned long size);
 
 void gcry_ac_io_init_writable_string(gcry_ac_io_t *ac_io,
-    unsigned char ** str, size_t * size);
+    unsigned char ** str, unsigned long * size);
 
 void gcry_ac_io_init_readable_callback(gcry_ac_io_t *ac_io,
     gcry_ac_data_read_cb_t cb, void * arg);
